+++ gordo/server/utils.py
@@ -6,14 +6,13 @@ import zlib
 import os
 import io
 import pickle
-import copy
 import re
 import shutil
 
 import dateutil
 import timeit
 from datetime import datetime
-from typing import Union, List, Any
+from typing import Union, List
 
 import pandas as pd
 import pyarrow as pa
@@ -485,33 +484,3 @@ def model_required(f):
             )
 
     return wrapper
-
-
-def find_path_in_dict(path: List[str], data: dict) -> Any:
-    """
-    Find a path in `dict` recursively
-
-    Examples
-    --------
-    >>> find_path_in_dict(["parent", "child"], {"parent": {"child": 42}})
-    42
-
-    Parameters
-    ----------
-    path: List[str]
-    data: dict
-
-    Returns
-    -------
-
-    """
-    reversed_path = copy.copy(path)
-    reversed_path.reverse()
-    curr_data = data
-    while len(reversed_path):
-        key = reversed_path.pop()
-        if key not in curr_data:
-            exception_path = ".".join(path[: len(path) - len(reversed_path)])
-            raise KeyError("'%s' is absent" % exception_path)
-        curr_data = curr_data[key]
-    return curr_data
